requirements.txt
dj-database-url==0.5.0
Django==3.0.7
django-cors-headers==3.5.0
django-heroku==0.3.1
django-storages==1.9.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
dropbox==10.2.0
gunicorn==20.0.4
Pillow==7.0.0
psycopg2==2.8.6
PyJWT==1.7.1
pytz==2019.3
whitenoise==5.1.0
